+++ src/html-renderer-svg.ts
